+++ src/lib.zig
@@ -40,6 +40,7 @@ pub const rules = struct {
     pub const net_status = @import("rules/net_status.zig");
     pub const prefetch = @import("rules/prefetch.zig");
     pub const runner = @import("rules/runner.zig");
+    pub const syntax = @import("rules/syntax.zig");
     pub const composite_steps = @import("rules/composite_steps.zig");
     pub const local_action = @import("rules/local_action.zig");
 };